+++ Changelog.md
@@ -5,6 +5,9 @@
 1. Add simplified ```readFile`` function.
    * [Pull request 347](https://bitbucket.org/osrf/sdformat/pull-requests/347)
 
+1. Remove boost::lexical cast instances
+   * [Pull request 342](https://bitbucket.org/osrf/sdformat/pull-requests/342)
+
 1. Remove boost regex and iostreams as dependencies
    * [Pull request 302](https://bitbucket.org/osrf/sdformat/pull-requests/302)